moved quiz.py and commands.py to bot_funcs and added updatedb.py back
